修正：パケットでのブロック座標をintで管理する

+++ src/main/java/nahama/ofalenmod/network/MApplyDetailedSetter.java
@@ -8,29 +8,29 @@ import nahama.ofalenmod.tileentity.TileEntityDetailedSetter;
 import net.minecraft.tileentity.TileEntity;
 
 public class MApplyDetailedSetter implements IMessage {
-	public short x, y, z;
+	public int x, y, z;
 
 	public MApplyDetailedSetter() {
 	}
 
 	public MApplyDetailedSetter(int x, int y, int z) {
-		this.x = (short) x;
-		this.y = (short) y;
-		this.z = (short) z;
+		this.x = x;
+		this.y = y;
+		this.z = z;
 	}
 
 	@Override
 	public void fromBytes(ByteBuf buf) {
-		x = buf.readShort();
-		y = buf.readShort();
-		z = buf.readShort();
+		x = buf.readInt();
+		y = buf.readInt();
+		z = buf.readInt();
 	}
 
 	@Override
 	public void toBytes(ByteBuf buf) {
-		buf.writeShort(x);
-		buf.writeShort(y);
-		buf.writeShort(z);
+		buf.writeInt(x);
+		buf.writeInt(y);
+		buf.writeInt(z);
 	}
 
 	public static class Handler implements IMessageHandler<MApplyDetailedSetter, IMessage> {
